renamed types and created local declarations
