+++ src/app/core/models/stats.model.ts
