+++ src/app/core/models/v2/stats-response.model.ts
@@ -35,7 +35,7 @@ export class StatsResponse {
     this.url = url;
   }
 
-  getDatum?(index: number) {
+  getDatum(index: number) {
     return this.data && this.data.length > 0 && this.data[index];
   }
 }
